fix: fix broken logout in secret bid auction example (#107)
Fixes the logout in the secret bid auction example, which is broken. In particular, after a logout and re-login, calls to the backend are still made with the old identity.

At the same time, the PR also

improves the example name on the main page
adds a code snipped that makes npm run dev work (which didn't work before)
reverses the order of the open and closes lots as they are displayed in the frontend, so that the most recent lots are on top