chore: CRP-2806 prepare new Motoko backend release (#147)
Fixes two broken links in the Motoko backend library README.

Also

sets the repository directly in the mops.toml rather than in the README. This also improves the "Package Quality" score on mops.one.
bumps the version of the Motoko backend library to 0.2.0 to be in line with the Rust backend library.