test: CRP-2752 run frontend tests concurrently (#186)
Pass `sequence.concurrent` to `vitest`. This instructs `vitest` to run
all tests concurrently and reduces the runtime of the current tests from
~60s to ~20s. The main reason of the longer runtime was that we made
quite a few update calls to pocket IC, which are not super fast. This
change is important to not get problems with CI when we add more tests.

I ran this 5 times locally and didn't observe any flakiness. If this
will be the case in the future, e.g., if we add more tests, then we
could either try to limit the maximum number of tests running
concurrently or change the strategy how we parallelize tests.