fix: dev dependencies in examples' frontends (#193)
The workflows for testing examples installed frontend dependencies from
the root of the repository, which is more than the dependencies of the
examples. Therefore, CI was green if an example was missing a dependency
but it was included in the root `package.json`. This PR fixes this logic
and adds missing dependencies.