chore: bump Rust backend deps (#213)
The main goal is to bump `ic-stable-structures` that introduce
performance improvements and some breaking changes. Why should we bump
it? Well, if we want to use the latest version of that crate in a
canister, we will also need to have it in `ic-vetkeys` because we need
it there for `KeyManager` and `EncryptedMaps`.

Also removes 2 unused deps in `ic-vetkeys`.