fix(crypto): basic_bls_signing example: use persistent actor and Motoko core (#233)
Adapts the basic_bls_signing example so that it works again with ICP
Ninja, which recently upgraded to DFX 0.29.1, which ships with a new
motoko compiler (0.16.1). The necessary changes for this were

* Declaring the actor as `persistent`
* Changing the `signatures` map to the stable mo:core/Map

As the examples was touched anyway, the PR switches from the outdated
`mo:base` to the new `mo:core` and fixes all compiler issues.

The Github workflow sets the `DFX_VERSION` environment variable, so that
only this example uses the new DFX version 0.29.1. This also required to
change the bind host from `localhost` to `127.0.0.1` as otherwise this
new DFX version hits the following error:
```
Error: The replica returned an HTTP Error: Http Error: status 400 Bad Request, content type "text/plain; charset=utf-8", content: error: no_authority
details:
The request is missing the required authority information (e.g. 'Host' header).
```

The changes were tested in ICP Ninja with the following link:


[![](https://icp.ninja/assets/open.svg)](http://icp.ninja/editor?g=https://github.com/dfinity/vetkeys/tree/franzstefan/CRP-2924-bls-signing/examples/basic_bls_signing/motoko)